pyproject.toml
[project]
name = "robottemplate"
version = "0.1.0"
description = "python automatik template"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "python-dotenv>=1.0.1",
  "pandas>=1.0.0",
  "os2apwrapper>=0.0.1",
  "topdeskpy>=0.1.0",
  "opuspy>=0.6.0",
  "ipykernel>=6.0.0"
]

[tool.uv.sources]
os2apwrapper = { git = "https://github.com/BorholmsRegionsKommuneIT/os2apwrapper" }
topdeskpy = { git = "https://github.com/BorholmsRegionsKommuneIT/topdeskpy" }
opuspy = { git = "https://github.com/BorholmsRegionsKommuneIT/opuspy" }

[tool.ruff]
line-length = 140